+++ __tests__/Broadcaster.spec.ts
@@ -32,9 +32,22 @@ describe("Broadcaster", () => {
     const observer2 = new Observer();
     const spy = vi.spyOn(observer, "onNotify");
     const spy2 = vi.spyOn(observer2, "onNotify");
-    broadcaster.subscribe([observer, observer2]);
+    broadcaster.subscribe(observer, observer2);
     broadcaster.notify(broadcaster, "destroy");
     expect(spy).toHaveBeenCalledOnce();
     expect(spy2).toHaveBeenCalledOnce();
+  });
+
+  test("can unsubscribe multiple observers", () => {
+    const broadcaster = new Broadcaster();
+    const observer = new Observer();
+    const observer2 = new Observer();
+    const spy = vi.spyOn(observer, "onNotify");
+    const spy2 = vi.spyOn(observer2, "onNotify");
+    broadcaster.subscribe(observer, observer2);
+    broadcaster.unsubscribe(observer, observer2);
+    broadcaster.notify(broadcaster, "destroy");
+    expect(spy).not.toHaveBeenCalled();
+    expect(spy2).not.toHaveBeenCalled();
   });
 });